[Hotfix] Re-arrange the test package structure of streampark-console-service.
